bugfix: fix resursive dns lookups, add label code to backend
this PR drastically overhauls the DNS code (such as making sure we get records from nameservers themselves, we support cname's better, we return the server responsible). I also added a VERY rough frontend implementation of this (mainly to make how the API works clear), but I will send this off to Ben before it is merged because the styling of it is incorrect.